Unifying version seperator.
Fixing issue where mod versions which used a simple ToString() could cause cultural issues in the version matching logic. (ex: fr-FR can cause a 1.0.0.0 to become a 1,0,0,0). This change unifies the separator using CulturalInfo to be '.' and then has the server change it back to the server's culture for printing to the screen.